webnfc-push/index.html

<!DOCTYPE html>
<!--
Copyright (c) 2017 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->
<meta charset=utf-8>
<title>Web NFC Test: push NFCMessage with 'text' recordType</title>
<link rel="author" title="Intel" href="http://www.intel.com"/>
<link rel="author" title="Zhao Ming" href="mailto:mingx.zhao@intel.com"/>
<link rel="help" href="https://w3c.github.io/web-nfc/#the-nfc-interface"/>
<script src="./resources/testharness.js"></script>
<script src="./resources/testharnessreport.js"></script>
<script src="./nfc_help.js"></script>
<meta name="flags" content="interact">
<meta name="timeout" content="long">

<div id="log"></div>

<h2>Precondition</h2>
<ol>
  <li>
   test
  </li>
</ol>


<script>

"use strict";

promise_test(t => {
  let watchOptions = { 
    data: [ 
        {recordType: "json", mediaType: "application/json", data: {name: "Benny Jensen", title: "Banker"}},
        {recordType: "json", mediaType: "application/json", data: {name: "Zoey Braun", title: "Engineer"}}
    ]
  }
  return navigator.nfc.push(watchOptions)
    .then(() => {
      return new Promise(resolve => {
        navigator.nfc.watch((message) => resolve(message), {url: document.baseURL, recordType: "json"})
      }).then((message) => {
        for (let record of message.data) {
          //assert_equals(record.data.title, watchOptions.data.title)
          //assert_equals(record.data.name, watchOptions.data.name)
          //assert_equals(record.data.meidaType, watchOptions.data.mediaType)
          //assert_equals(record.data.recordType, watchOptions.data.recordType)
          console.log("000")
          console.log(record)
          console.log("15151515")
        }
        for (let options of watchOptions.data) {
          console.log("1111")
          console.log(options)
          console.log("222")
          console.log(record)
        }
      })
    })
}, "Push and Read JSON 'serialized and deserialized'.")

function testNFCWatchOptions(args) {

}

</script>
